apply SLA schema, workplace: add search from FB and handle workplace not found
