+++ imports/ui/pages/index.js
@@ -3,8 +3,7 @@ export {default as BlankPage} from './Blank';
 
 // Main Pages
 export {default as HomePage} from './Home';
-export {default as WorkplacesPage} from './Workplaces.new';
-// export {default as WorkplacesPage} from './Workplaces';
+export {default as WorkplacesPage} from './Workplaces';
 export {default as SLAsPage} from './SLAs';
 export {default as AccessListPage} from './AccessList';
